research_framework.py
import os
import yaml
import torch
import torch.nn as nn
from PIL import Image
import numpy as np
from transformers import (
    BlipProcessor, BlipForConditionalGeneration,
    LlavaNextProcessor, LlavaNextForConditionalGeneration,
    AutoProcessor, AutoModelForVision2Seq
)
import albumentations as A
from sentence_transformers import SentenceTransformer, util
import wandb
from typing import List, Dict, Any, Type

# ==========================================
# 1. Advanced Model Registry (R&D Logic)
# ==========================================
class LMMRunner:
    """Modular wrapper for state-of-the-art Large Multimodal Models (LMMs)."""
    def __init__(self, model_id: str, device: str = "cuda"):
        self.device = device
        self.model_id = model_id
        
        # Determine model architecture automatically
        if "llava" in model_id.lower():
            self.processor = LlavaNextProcessor.from_pretrained(model_id)
            self.model = LlavaNextForConditionalGeneration.from_pretrained(
                model_id, torch_dtype=torch.float16, low_cpu_mem_usage=True
            ).to(device)
        else: # Default to BLIP/InstructBLIP family
            self.processor = BlipProcessor.from_pretrained(model_id)
            self.model = BlipForConditionalGeneration.from_pretrained(
                model_id, torch_dtype=torch.float16
            ).to(device)

    def generate(self, image: Image.Image, prompt: str = None) -> str:
        if prompt:
            # LMMs like LLaVA require specific formatting
            inputs = self.processor(text=f"USER: <image>\n{prompt} ASSISTANT:", images=image, return_tensors="pt").to(self.device)
        else:
            inputs = self.processor(images=image, return_tensors="pt").to(self.device)
            
        with torch.no_grad():
            output = self.model.generate(**inputs, max_new_tokens=100)
        return self.processor.batch_decode(output, skip_special_tokens=True)[0]

# ==========================================
# 2. Advanced Adversarial & Stochastic Noise
# ==========================================
class PerturbationEngine:
    """Generates natural and adversarial visual corridors."""
    def __init__(self, config: Dict):
        self.config = config
        
    def apply_stochastic(self, img_np: np.ndarray, n_type: str, severity: float) -> np.ndarray:
        """Albumentations-based natural corruptions."""
        if n_type == "gaussian":
            transform = A.GaussNoise(var_limit=(severity*0.5, severity*0.5), p=1.0)
        elif n_type == "blur":
            k = int(severity * 25) + 1
            if k % 2 == 0: k += 1
            transform = A.GaussianBlur(blur_limit=(k, k), p=1.0)
        elif n_type == "jpeg":
            q = max(1, 100 - int(severity * 99))
            transform = A.ImageCompression(quality_lower=q, quality_upper=q, p=1.0)
        else:
            return img_np
        return transform(image=img_np)["image"]

    def apply_adversarial_pgd(self, model: nn.Module, inputs: Dict, epsilon: float = 0.03, alpha: float = 0.005, iters: int = 20):
        """[RESEARCH MODULE] PGD attack on visual embedding space (requires manual torch logic)."""
        # (This is a simplified research skeleton for adversarial probing)
        pass 

# ==========================================
# 3. Evaluation Framework (Metric Tier)
# ==========================================
class RobustnessEvaluator:
    def __init__(self, embedder_id: str = "all-mpnet-base-v2"):
        self.embedder = SentenceTransformer(embedder_id)

    def compute_semantic_drift(self, ground_truth: str, generated: str) -> float:
        embs = self.embedder.encode([ground_truth, generated])
        return 1.0 - util.cos_sim(embs[0], embs[1]).item()

    def chairs_score(self, caption: str, ground_truth_objects: List[str]) -> bool:
        """[R&D METRIC] Simplified CHAIR-i: Does caption contain objects not in ground truth?"""
        # Implement object extraction with Spacy/NLTK for research rigor
        pass

# ==========================================
# 4. Main Experiment Orchestrator
# ==========================================
def main_research_loop(config_path: str):
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)

    # Initialize WandB for experiment tracking (Industry Secret Sauce)
    if config.get("use_wandb", False):
        wandb.init(project="VAlign-Robust", config=config)

    model_runner = LMMRunner(config["model_id"])
    engine = PerturbationEngine(config)
    evaluator = RobustnessEvaluator()

    # (Dataset loop logic would go here)
    # 1. Load image
    # 2. Generate clean baseline
    # 3. Loop types x intensities
    # 4. Compute metrics
    # 5. Log to W&B
    
    print("[SUCCESS] Research Experiment Initialized & Configured.")

if __name__ == "__main__":
    # Example config generation for the user
    example_config = {
        "model_id": "llava-hf/llava-1.5-7b-hf",
        "noise_types": ["gaussian", "blur", "jpeg"],
        "intensity_steps": [0.0, 0.2, 0.4, 0.6, 0.8, 1.0],
        "use_wandb": False
    }
    with open("experiment_config.yaml", "w") as f:
        yaml.dump(example_config, f)
    
    print("Run with: python research_framework.py")
